Fix bug when exiting scilab. Fix bug with swap_handles unitary test.

+++ scilab/modules/renderer/src/cpp/DrawableObjectJoGL.cpp
@@ -54,9 +54,12 @@ void DrawableObjectJoGL::show( void )
 /*------------------------------------------------------------------------------------------*/
 void DrawableObjectJoGL::destroy( void )
 {
-  getJavaMapper()->destroy(sciGetNum(sciGetParentFigure(getDrawer()->getDrawedObject())));
-  delete getJavaMapper();
-  m_pJavaMapper = NULL;
+  if (m_pJavaMapper != NULL)
+  {
+    getJavaMapper()->destroy(sciGetNum(sciGetParentFigure(getDrawer()->getDrawedObject())));
+    delete getJavaMapper();
+    setJavaMapper(NULL);
+  }
 }
 /*------------------------------------------------------------------------------------------*/
 }
